requirements.txt
fastapi
uvicorn
apscheduler
PyYAML
pyinstaller